requirements.txt
# dev
pytest
expects

# prod
requests
google-api-python-client
google-auth